+++ include/Parts.hpp
@@ -100,14 +100,14 @@ public:
 	 *
 	 * @return the number of parts
 	 */
-	unsigned int nparts(void) const { return filterid_->size(); }
+	size_t nparts(void) const { return filterid_->size(); }
 
 	/*! @brief get the number of mixtures for the current part
 	 *
 	 * @return the number of mixtures. Note that not all components
 	 * and parts will have the same number of mixtures
 	 */
-	unsigned int nmixtures(void) const { return (*filterid_)[self_].size(); }
+	size_t nmixtures(void) const { return (*filterid_)[self_].size(); }
 
 	/*! @brief the current flattened tree index
 	 *
@@ -121,7 +121,7 @@ public:
 	 * @param mixture the part mixture of interest
 	 * @return the associated filter
 	 */
-	const cv::Mat& filter(unsigned int mixture = 0) const {
+	const cv::Mat& filter(size_t mixture = 0) const {
 		assert((*filterid_)[self_].size() > mixture);
 		return (*filtersw_)[(*filterid_)[self_][mixture]];
 	}
@@ -132,7 +132,7 @@ public:
 	 */
 	void filters(vectorMat& out) const {
 		out.clear();
-		for (unsigned int m = 0; m < filterid_[self_].size(); ++m) {
+		for (size_t m = 0; m < filterid_[self_].size(); ++m) {
 			out.push_back((*filtersw_)[(*filterid_)[self_][m]]);
 		}
 	}
@@ -147,7 +147,7 @@ public:
 	 */
 	std::vector<ComponentPart> children(void) const {
 		std::vector<ComponentPart> c;
-		for (unsigned int n = self_; (*parentid_)[n] <= self_; ++n) {
+		for (size_t n = self_; (*parentid_)[n] <= self_; ++n) {
 			ComponentPart cp(*this, n);
 			if ((*parentid_)[n] == self_) c.push_back(cp);
 		}
@@ -162,29 +162,29 @@ public:
 	 * @param mixture the part mixture to retrieve
 	 * @return the associated score for this part's mixture
 	 */
-	cv::Mat& score(vectorMat& scores, unsigned int mixture = 0) const {
+	cv::Mat& score(vectorMat& scores, size_t mixture = 0) const {
 		assert((*filterid_)[self_].size() > mixture);
 		return scores[(*filterid_)[self_][mixture]];
 	}
 	//! the part's filter index
-	int filteri(unsigned int mixture = 0) const { return (*filtersi_)[(*filterid_)[self_][mixture]]; }
+	int filteri(size_t mixture = 0) const { return (*filtersi_)[(*filterid_)[self_][mixture]]; }
 	//! the part's bias
-	vectorf bias(unsigned int mixture = 0) const {
+	vectorf bias(size_t mixture = 0) const {
 		const int offset = (*biasid_)[self_][mixture];
 		return vectorf(&((*biasw_)[offset]), &((*biasw_)[offset+nmixtures()]));
 	}
 	//! the part's bias index
-	int biasi(unsigned int mixture = 0) const { return (*biasi_)[(*biasid_)[self_][mixture]]; }
+	int biasi(size_t mixture = 0) const { return (*biasi_)[(*biasid_)[self_][mixture]]; }
 	//! the part's deformation weights
-	vectorf defw(unsigned int mixture = 0) const { return (*defw_)[(*defid_)[self_][mixture]]; }
+	vectorf defw(size_t mixture = 0) const { return (*defw_)[(*defid_)[self_][mixture]]; }
 	//! the part's deformation indices
-	int defi(unsigned int mixture = 0) const { return (*defi_)[(*defid_)[self_][mixture]]; }
+	int defi(size_t mixture = 0) const { return (*defi_)[(*defid_)[self_][mixture]]; }
 	//! the part's anchor (relative to its parent part)
-	cv::Point anchor(unsigned int mixture = 0) const { return (*anchors_)[(*defid_)[self_][mixture]]; }
+	cv::Point anchor(size_t mixture = 0) const { return (*anchors_)[(*defid_)[self_][mixture]]; }
 	//! the x size (width) of the part
-	unsigned int xsize(unsigned int mixture = 0) const { return (*filtersw_)[(*filterid_)[self_][mixture]].rows; }
+	size_t xsize(size_t mixture = 0) const { return (*filtersw_)[(*filterid_)[self_][mixture]].rows; }
 	//! the y size (height) of the part
-	unsigned int ysize(unsigned int mixture = 0) const { return (*filtersw_)[(*filterid_)[self_][mixture]].rows; }
+	size_t ysize(size_t mixture = 0) const { return (*filtersw_)[(*filterid_)[self_][mixture]].rows; }
 	//! is the current part the root
 	bool isRoot(void) const { return self_ == 0; }
 };
@@ -241,18 +241,18 @@ public:
 	 * @param p the part to reference within that component (defaults to the root)
 	 * @return the ComponentPart for component c at node p
 	 */
-	ComponentPart component(unsigned int c, unsigned int p = 0) {
+	ComponentPart component(size_t c, size_t p = 0) {
 		assert(c < biasid_.size() && c < filterid_.size() && c < parentid_.size());
 		return ComponentPart(filtersw_, filtersi_, biasw_, biasi_, anchors_, defw_, defi_, defid_[c], biasid_[c], filterid_[c], parentid_[c], p);
 	}
 	//! the number of components in the model
-	unsigned int ncomponents(void) const { return filterid_.size(); }
+	size_t ncomponents(void) const { return filterid_.size(); }
 	/*! @brief the number of parts within a component
 	 *
 	 * @param c the component of interest
 	 * @return the number of parts
 	 */
-	unsigned int nparts(unsigned int c) const {
+	size_t nparts(size_t c) const {
 		assert(c < biasid_.size() && c < filterid_.size() && c < parentid_.size());
 		return filterid_[c].size();
 	}
